fix: Controller division
